KA-100 Rename itemCreateFactory to match export
