+++ onboarding-task/test/reducers/itemReducer.test.js
@@ -2,7 +2,7 @@ import {
   createItem,
   updateItem,
 } from '../../src/actions/actionCreators.js';
-import { createItemFactory } from '../../src/actions/itemCreateFactory.js';
+import { createItemFactory } from '../../src/actions/createItemFactory.js';
 import { Item } from '../../src/models/Item.js';
 import { itemReducer } from '../../src/reducers/itemsReducers/itemReducer.js';
